requirements.txt
python-jobspy
playwright
playwright-stealth
groq
python-docx
reportlab
supabase
apscheduler
python-dotenv
pandas
